third_party/blink/web_tests/fast/css/getComputedStyle/computed-style-mask-box-image.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<p>Test computed style for the -webkit-mask-box-image property and sub-properties</p>
<div id="console"></div>
<script>

var testDiv = document.createElement('div');
document.body.appendChild(testDiv);

function computedMaskBoxImageStyle(maskBoxStyle, property) {
    testDiv.style.webkitMaskBoxImage = 'none';
    testDiv.style.webkitMaskBoxImage = maskBoxStyle;
    return window.getComputedStyle(testDiv).getPropertyValue(property);
}

shouldBeEqualToString("computedMaskBoxImageStyle('linear-gradient(1turn, red calc(20% * 2 / 4), blue)', '-webkit-mask-box-image-source')", 'linear-gradient(360deg, rgb(255, 0, 0) 10%, rgb(0, 0, 255))');
shouldBeEqualToString("computedMaskBoxImageStyle('radial-gradient(ellipse 60% 40% at calc(10% + 2em) 20%, red, blue)', '-webkit-mask-box-image-source')", 'radial-gradient(60% 40% at calc(10% + 32px) 20%, rgb(255, 0, 0), rgb(0, 0, 255))');
shouldBeEqualToString("computedMaskBoxImageStyle('conic-gradient(from 1.5708rad, red 0deg, blue calc(180deg - 10deg))', '-webkit-mask-box-image-source')", 'conic-gradient(from 90.0002deg, rgb(255, 0, 0) 0deg, rgb(0, 0, 255) 170deg)');

</script>
</body>
</html>
